+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixes
+- It's no longer possible to forcifully enable transformer not supported in installed Robot Framework version ([#281](https://github.com/MarketSquare/robotframework-tidy/issues/281), [#283](https://github.com/MarketSquare/robotframework-tidy/issues/283))
+
 ### Other
 - You can now disable coloring the output with ``--no-color`` cli option or by setting ``$NO_COLOR`` environment variable ([#268](https://github.com/MarketSquare/robotframework-tidy/issues/268))